backend/requirements.txt
Flask
Flask-CORS
mysql-connector-python
python-keycloak
PyYAML
gunicorn
